rag_agent/requirements.txt
llama-index
llama-index-llms-ollama
llama-index-embeddings-huggingface
requests
python-dotenv
